fix(147): F124 require consumer audience on pending-applications
The /api/v1/wallet/pending-applications group used plain .RequireAuthorization(),
so a platform token could read/set a citizen's notice. Gate it with
RequireConsumerAudience, matching every sibling citizen surface.

Also extracts the endpoint-metadata inspection into a shared
EndpointAuthorizationMetadata harness (permissive IServiceProviderIsService) and
refactors the US1 system-wallet metadata test to use it; adds
PendingApplicationAuthorizationTests. Whole Wallet suite green (852).
